refactor: rework merge action structure
