+++ setup.cfg
@@ -73,7 +73,7 @@ mergify_actions =
     post_check = mergify_engine.actions.post_check:PostCheckAction
     label = mergify_engine.actions.label:LabelAction
     delete_head_branch = mergify_engine.actions.delete_head_branch:DeleteHeadBranchAction
-    merge = mergify_engine.actions.merge.action:MergeAction
+    merge = mergify_engine.actions.merge:MergeAction
     close = mergify_engine.actions.close:CloseAction
     comment = mergify_engine.actions.comment:CommentAction
     review = mergify_engine.actions.review:ReviewAction
